+++ packages/logger/eslint.config.mjs
@@ -1,0 +1,5 @@
+import { defineScTsConfig } from '@shiftcode/eslint-config-recommended'
+
+export default defineScTsConfig({
+  languageOptions: { parserOptions: { project: ['./tsconfig.json', './tsconfig.jest.json'] } },
+})