+++ packages/eslint-plugin-rules/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.1](https://github.com/shiftcode/sc-commons-public/compare/@shiftcode/eslint-plugin-rules@4.0.0...@shiftcode/eslint-plugin-rules@4.0.1) (2025-09-09)
+
+**Note:** Version bump only for package @shiftcode/eslint-plugin-rules
+
+# [4.0.0](https://github.com/shiftcode/sc-commons-public/compare/@shiftcode/eslint-plugin-rules@3.0.0...@shiftcode/eslint-plugin-rules@4.0.0) (2025-05-15)
+
+### Features
+
+- **package:** specify node version in the individual packages ([78e01d0](https://github.com/shiftcode/sc-commons-public/commit/78e01d0be016e22584a17e7c021cc1b4408c4d1f))
+
+### BREAKING CHANGES
+
+- **package:** Requires Node 20 or >=22
+
+# [3.0.0](https://github.com/shiftcode/sc-commons-public/compare/@shiftcode/eslint-plugin-rules@2.0.1...@shiftcode/eslint-plugin-rules@3.0.0) (2025-01-03)
+
+### Build System
+
+- **deps:** upgrade @angular-eslint/\* ([fb1b0b4](https://github.com/shiftcode/sc-commons-public/commit/fb1b0b4161aaf9e23345ed543f598a5a4a79f7ba))
+
+### BREAKING CHANGES
+
+- **deps:** Requires angular 19.
+
 ## [2.0.1](https://github.com/shiftcode/sc-commons-public/compare/@shiftcode/eslint-plugin-rules@2.0.0...@shiftcode/eslint-plugin-rules@2.0.1) (2024-07-13)
 
 **Note:** Version bump only for package @shiftcode/eslint-plugin-rules