+++ packages/publish-helper/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.2](https://github.com/shiftcode/sc-commons-public/compare/@shiftcode/publish-helper@4.1.1...@shiftcode/publish-helper@4.1.2) (2025-09-11)
+
+**Note:** Version bump only for package @shiftcode/publish-helper
+
+## [4.1.1](https://github.com/shiftcode/sc-commons-public/compare/@shiftcode/publish-helper@4.1.0...@shiftcode/publish-helper@4.1.1) (2025-09-09)
+
+**Note:** Version bump only for package @shiftcode/publish-helper
+
+# [4.1.0](https://github.com/shiftcode/sc-commons-public/compare/@shiftcode/publish-helper@4.0.0...@shiftcode/publish-helper@4.1.0) (2025-09-04)
+
+### Features
+
+- **publish-helper:** support latest branch-utilities to allow github copilot branch names ([6777213](https://github.com/shiftcode/sc-commons-public/commit/67772138cc2070ed5f5cd8ce8630dbd9aadfbbfc))
+
+# [4.0.0](https://github.com/shiftcode/sc-commons-public/compare/@shiftcode/publish-helper@3.1.0...@shiftcode/publish-helper@4.0.0) (2025-05-15)
+
+### Features
+
+- **package:** specify node version in the individual packages ([78e01d0](https://github.com/shiftcode/sc-commons-public/commit/78e01d0be016e22584a17e7c021cc1b4408c4d1f))
+
+### BREAKING CHANGES
+
+- **package:** Requires Node 20 or >=22
+
+# [3.1.0](https://github.com/shiftcode/sc-commons-public/compare/@shiftcode/publish-helper@3.0.1...@shiftcode/publish-helper@3.1.0) (2025-05-14)
+
+### Features
+
+- **publish-helper:** update conventional-changelog-angular ([1ec776b](https://github.com/shiftcode/sc-commons-public/commit/1ec776b3f0017dcad90e92376afcdd11ab77e418))
+
 ## [3.0.1](https://github.com/shiftcode/sc-commons-public/compare/@shiftcode/publish-helper@3.0.0...@shiftcode/publish-helper@3.0.1) (2024-07-26)
 
 **Note:** Version bump only for package @shiftcode/publish-helper